+++ grpcclient/grpcclient.go
@@ -102,7 +102,7 @@ func New(cfg Config, opts ...Option) *Component {
 	c := &Component{
 		cfg:    cfg,
 		log:    nopLogger{},
-		name:   "grpc-client",
+		name:   "grpcclient",
 		stopCh: make(chan struct{}), // initialised so Stop-before-Start is safe
 	}
 	for _, o := range opts {